requirements.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
Markdown
Pillow
pytest
python-dotenv
crispy-bootstrap5
coreapi
django-celery-beat
django-crispy-forms
django-debug-toolbar
django-extensions
psycopg2
celery
PyJWT
epicbox